+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/TypeCoercionHelper.scala
@@ -184,11 +184,11 @@ abstract class TypeCoercionHelper {
   protected def findWiderTypeForDecimal(dt1: DataType, dt2: DataType): Option[DataType] = {
     (dt1, dt2) match {
       case (t1: DecimalType, t2: DecimalType) =>
-        Some(DecimalPrecision.widerDecimalType(t1, t2))
+        Some(DecimalPrecisionTypeCoercion.widerDecimalType(t1, t2))
       case (t: IntegralType, d: DecimalType) =>
-        Some(DecimalPrecision.widerDecimalType(DecimalType.forType(t), d))
+        Some(DecimalPrecisionTypeCoercion.widerDecimalType(DecimalType.forType(t), d))
       case (d: DecimalType, t: IntegralType) =>
-        Some(DecimalPrecision.widerDecimalType(DecimalType.forType(t), d))
+        Some(DecimalPrecisionTypeCoercion.widerDecimalType(DecimalType.forType(t), d))
       case (_: FractionalType, _: DecimalType) | (_: DecimalType, _: FractionalType) =>
         Some(DoubleType)
       case _ => None
